Se refactoriza jwt en user para que se pueda usar en otras rutas
